update docs structure for tooltip API spec
