+++ packages/components/src/__overlays__/Tooltip/v3/_docs/Tooltip.mdx
@@ -1,15 +1,24 @@
-import { Meta, Canvas, Controls  } from "@storybook/blocks"
-import { ResourceLinks, Installation, SbText } from "~storybook/components"
-import * as TooltipStories from "../Tooltip.spec.stories"
+import { Canvas, Meta, Unstyled, Controls } from "@storybook/blocks"
+import { CheckIcon, ClearIcon } from "~components/Icon"
+import { ResourceLinks, SbText, SbHeading, Installation } from "~storybook/components"
+import * as TooltipStories from "./Tooltip.stories"
+import AnatomyPng from "./assets/tooltip_anatomy.png"
+import VariantsPng from "./assets/tooltip_variants.png"
+import WhenToUsePng from "./assets/tooltip_when_to_use.png"
 
-<Meta of={TooltipStories} />
+<Meta title="Overlays/Tooltip/v3/Usage Guidelines" />
 
 # Tooltip (v3)
+<SbText variant="extra-small">
+  Updated June X, 2024
+</SbText>
+
 
 <ResourceLinks
   sourceCode="https://github.com/cultureamp/kaizen-design-system/tree/main/packages/components/src/Tooltip"
   figma=""
   designGuidelines="/?path=/docs/overlays-tooltip-v3-usage-guidelines--docs"
+  apiSpecification="/?path=/docs/overlays-tooltip-v3-api-specifications--docs"
   className="!mb-8"
 />
 
@@ -26,11 +35,121 @@ import * as TooltipStories from "../Tooltip.spec.stories"
 </SbText>
 
 <Canvas of={TooltipStories.Playground} />
-<Controls of={TooltipStories.Playground} include={["placement", "isOpen"]} />
+<Controls of={TooltipStories.Playground} include={["placement", "isOpen"]} className="mb-64" />
 
-## Quick links
 
-<SbText>
-  - See our design documentation in the [Usage Guidelines](/?path=/docs/overlays-tooltip-v3-usage-guidelines--docs)
-  - See our engineering documentation in the [API Specification](/?path=/docs/overlays-tooltip-v3-api-specifications--docs)
-</SbText>
+### Anatomy
+
+<div className="mb-32">
+  <SbText>
+    - The anatomy of a tooltip is made up of a container, description and arrow.
+      - **They do not include a title or heading**. If you need to include a title, use a Popover
+    - A tooltip is connected to a trigger element, such as a button or icon, that shows the tooltip when the interactive trigger element is hovered or focused. (More details below.)
+  </ SbText>
+</div>
+
+<img src={AnatomyPng} alt="Indicates the anatomy of the Tooltip component, the Container, Label, zArrow and Trigger" className="mb-64" />
+
+### When to use
+
+<div className="mt-16 mb-32">
+  <SbText>
+    - Tooltips are used to provide additional information when space is tight.
+    - Tooltips must be concise with one or more lines of text.
+    - Tooltips must be accessible via hover and focus only.
+    - Trigger tooltips from Interactive elements (Buttons, Links, Icon buttons) or Non-interactive elements (Icons, Abbreviations Truncated text)—be mindful of keyboard accessibility
+    - **Never include any kind of error messages** in a tooltip
+  </ SbText>
+</ div>
+
+<img src={WhenToUsePng} alt="Illustration of when to use the Tooltip component" className="mb-32"/>
+
+### Dos and don'ts
+<Unstyled>
+  <div className="flex flex-col mt-16 gap-40 mb-64">
+    <div className="flex flex-col m-0 gap-16">
+        <SbHeading variant="heading-4">
+          #### Do use tooltips to describe icons
+        </SbHeading>
+        <SbText variant="body" tag="span">
+          Icons are not always easy to identify on their own. When you use components that don&apos;t have labels — for example, icon-only action buttons and tabs — make sure to use tooltips to provide context for the icons.
+        </SbText>
+      <div className="grid grid-cols-2 gap-16">
+        <div className="relative [&>*]:m-0">
+          <div className="bg-green-500 absolute top-6 left-6 z-10 py-4 px-6 font-bold text-white rounded flex gap-2 items-center justify-center">
+            <CheckIcon role="presentation" />
+            <SbText color="white">
+              <strong>Do</strong>
+            </SbText>
+          </div>
+          <Canvas of={TooltipStories.Primary} className="border-green-500 !rounded" />
+        </div>
+      </div>
+    </div>
+
+    <div className="flex flex-col gap-16">
+        <SbHeading variant="heading-4">
+          #### Don&apos;t use tooltips to communicate crucial information
+        </SbHeading>
+        <SbText>
+          Tooltips should never stop a user (or be a gate) from completing a task or performing an action. They disappear so they should never contain critical information that a user would absolutely need to proceed, or information without which a user may cause errors. If the information is crucial for a user to proceed, consider using a modal or inline notification instead.
+        </SbText>
+      <div className="grid grid-cols-2 gap-16">
+        <div className="relative [&>*]:m-0">
+          <div className="bg-green-500 absolute top-6 left-6 z-10 py-4 px-6 font-bold text-white rounded flex gap-2 items-center justify-center">
+            <CheckIcon role="presentation" />
+            <SbText color="white"><strong>Do</strong></SbText>
+          </div>
+          <Canvas of={TooltipStories.DoFieldTooltip} className="border-green-500 !rounded-[10px]" />
+        </div>
+        <div className="relative [&>*]:m-0">
+          <div className="bg-gray-500 absolute top-6 left-6 z-10 py-4 px-6 font-bold text-white rounded flex gap-2 items-center justify-center">
+            <ClearIcon role="presentation" />
+            <SbText color="white"><strong>Don&apos;t</strong></SbText>
+          </div>
+          <Canvas of={TooltipStories.DontFieldTooltip} className="border-gray-500 !rounded-[10px]" />
+        </div>
+      </div>
+    </div>
+
+    <div className="flex flex-col m-0 gap-16">
+        <SbHeading variant="heading-4">
+          #### Do be concise
+        </SbHeading>
+        <SbText>
+          Good tooltips contain concise and helpful information, keep it short. If you need more space, consider using a Popover. Don’t use it to duplicate information that already exists on the page. Instead, closely consider its placement and whether the disclosed copy will inform the action the user is about to perform.
+        </SbText>
+      <div className="grid grid-cols-2 gap-16">
+        <div className="relative [&>*]:m-0">
+          <div className="bg-green-500 absolute top-6 left-6 z-10 py-4 px-6 font-bold text-white rounded flex gap-2 items-center justify-center">
+            <CheckIcon role="presentation" />
+            <SbText color="white"><strong>Do</strong></SbText>
+          </div>
+          <Canvas of={TooltipStories.DoConcise} className="border-green-500 !rounded-[10px]" />
+        </div>
+        <div className="relative [&>*]:m-0">
+          <div className="bg-gray-500 absolute top-6 left-6 z-10 py-4 px-6 font-bold text-white rounded flex gap-2 items-center justify-center">
+            <ClearIcon role="presentation" />
+            <SbText color="white"><strong>Don&apos;t</strong></SbText>
+          </div>
+          <Canvas of={TooltipStories.DontConcise} className="border-gray-500 !rounded-[10px]" />
+        </div>
+      </div>
+    </div>
+  </div>
+</Unstyled>
+
+### Variations
+
+<img src={VariantsPng} alt="Indicates the anatomy of the Tooltip component, the Container, Label, zArrow and Trigger" className="my-16" />
+
+<div className="mb-64">
+  <SbText>
+    - Type: Dropdown to choose from Default (dark tooltip) and Reverse (light tooltip)
+    - Direction: Dropdown to choose where the arrow points
+    - Tooltip label: Input field allows you to type in the label from the properties panel
+    - Use strong signifiers for trigger elements to ensure the tooltip is discoverable.
+    - The proximity of tooltips is always paired to the element with which they are associated.
+  </SbText>
+</ div>
+
