+++ storybook/components/SbText/SbText.tsx
@@ -2,7 +2,7 @@ import React from "react"
 import { Unstyled } from "@storybook/blocks"
 import { Text, TextProps } from "~components/Text"
 
-/** A sensible defaulft and convenience wrapper for Storybook text so we don't have to wrap it in an unstyled all the time */
+/** A sensible default and convenience wrapper for Storybook text so we don't have to wrap it in an unstyled all the time */
 export const SbText = ({
   tag = "span",
   variant = "body",
