+++ storybook/components/index.ts
@@ -6,3 +6,4 @@ export * from "./ResourceLinks"
 export * from "./SbReleaseNotification"
 export * from "./TailwindStoryTemplate"
 export * from "./SbText"
+export * from "./SbHeading"
